test: pulling protocol specific test code into protocol-specific test files
